requirements.txt
groq>=0.4.1
ollama>=0.1.7
python-dotenv==1.0.0
requests>=2.28.0
pandas>=1.5.0
matplotlib>=3.6.0
numpy>=1.24.0
psycopg2-binary>=2.9.9
alembic>=1.12.1
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
WTForms==3.0.1
SQLAlchemy==2.0.21
Werkzeug==2.3.7
Jinja2==3.1.2
email-validator==2.0.0
bcrypt==4.0.1
gunicorn==21.2.0
celery>=5.3.0